Clarify Pareto smoke output complexity
Constraint: Pareto complexity is a deterministic integer work score, not a floating timing metric.

Rejected: Leaving the smoke script as a plain ad-hoc benchmark | it obscures that the entrypoint is only a smoke/provenance check.

Confidence: high

Scope-risk: narrow

Directive: Keep Pareto benchmark artifacts explicit about smoke status until real result matrices are designed.

Tested: .venv/bin/python -m pytest tests/test_kernel_pareto.py -q; .venv/bin/ruff check veqpy/kernels/pareto.py tests/test_kernel_pareto.py benchmarks/numba_pareto.py; .venv/bin/python benchmarks/numba_pareto.py --strategy adaptive --metric rms --pareto-by counts --max-candidates 4 --no-write; .venv/bin/python -m compileall -q veqpy/kernels/pareto.py tests/test_kernel_pareto.py benchmarks/numba_pareto.py

Not-tested: Full persisted benchmark matrix generation.

+++ veqpy/kernels/pareto.py
@@ -66,7 +66,7 @@ class ParetoSample:
     signature: KernelParetoSignature
     counts: int
     time: float
-    complexity: float
+    complexity: int
     shape_error: float
     result: SolveResult
 
@@ -123,10 +123,10 @@ def normalize_shape_error_thresholds(
     )
 
 
-def pareto_sample_complexity(result: SolveResult, topology: KernelTopology) -> float:
+def pareto_sample_complexity(result: SolveResult, topology: KernelTopology) -> int:
     nx = int(topology.x_size)
     nx2 = nx * nx
-    return float(
+    return (
         int(result.nfev) * nx
         + int(result.jvp_evaluations) * nx2
         + int(result.jacobian_component_evaluations) * nx2
